+++ CJBCheatsMenu/docs/release-notes.md
@@ -8,6 +8,7 @@
 * You can now [make personal edits to the warp list](README.md#personal-warps).
 * You can now set tomorrow's weather to green rain.
 * The time slider can now be set at 10-minute intervals instead of 1-hour intervals.
+* The "time frozen" message now fades out after a few seconds in a location, to reduce conflicts with other mods' UIs in the top-left corner.
 * Fixed 'freeze time' and 'grow' keybinds still triggered when typing into a text field.
 * Fixed missing translations for default tab names in Generic Mod Config Menu UI.
 * Fixed unknown weathers shown as blank in the weather tab; it'll now show the weather ID instead.